[MCU8BIT][NUVOTON] update i2c, add i2c device scan
